requirements.txt
pynput
pyperclip
pyautogui
psutil
pygetwindow
cryptography
pillow
